Fix potential overflow error in CEQs (XCC)
The issue was that a consumer that came in during overflow recovery could
see that there was no overflow and return FALSE, meaning the CEQ was empty,
even though there were older messages.

Consider, the kernel already posted two messages, set overflow, and the
ring is empty:

Thread 1                      Thread 2
--------                      --------
see empty ring                see empty ring
see overflow is on
grab lock
clear overflow
extract a message
                              sees overflow is off
                              returns FALSE
sets overflow
unlocks
returns TRUE

And there's still a message in the CEQ that thread 2 should have grabbed.

While doing this change, I also changed nr_events to an unsigned.  That was
my original intent (based on the usage in epoll), and making the change now
keeps this commit from changing the size of the CEQ, which keeps everyone
from having to rebuild every application.

Reinstall your kernel headers.

+++ user/parlib/include/parlib/ceq.h
@@ -31,7 +31,8 @@ __BEGIN_DECLS
  * probably want to do it yourself. */
 #define CEQ_DEFAULT_SZ 128
 
-void ceq_init(struct ceq *ceq, uint8_t op, size_t nr_events, size_t ring_sz);
+void ceq_init(struct ceq *ceq, uint8_t op, unsigned int nr_events,
+              size_t ring_sz);
 bool get_ceq_msg(struct ceq *ceq, struct event_msg *msg);
 bool ceq_is_empty(struct ceq *ceq);
 void ceq_cleanup(struct ceq *ceq);
